% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_GUI.R
\name{combine_netWprop}
\alias{combine_netWprop}
\title{Combine network data with propagation values}
\usage{
combine_netWprop(net, propag, totaldataframe)
}
\arguments{
\item{net}{igraph object}

\item{propag}{propagation matrix}

\item{totaldataframe}{metainformation matrix}
}
\description{
Support function for the network visualization
}
