% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_RWR.R
\name{complete_m}
\alias{complete_m}
\title{Harmonizes the matrix of sample's profiles with respect the network}
\usage{
complete_m(m_profiles, net)
}
\arguments{
\item{m_profiles}{Profile matrix with rows matching nodes of the network}

\item{net}{Adjacency matrix of a network s.t. nodes have same names of the rows in the profile matrix}
}
\value{
The profile matrix without the rows not having a match with a node of the network
}
\description{
Remove all the genes and fragments that don't match between the sample's profiles and the nodes inside the network
}
