% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_GUI.R
\name{create_net2plot}
\alias{create_net2plot}
\title{Create an igraph network object for visualization purposes}
\usage{
create_net2plot(
  g_net,
  input_m,
  gf_prop,
  ann_net_b,
  frag_pattern = "F",
  ff_net = NULL,
  ff_prop = NULL
)
}
\arguments{
\item{g_net}{edge list matrix of the gene gene interaction network}

\item{input_m}{sample's profile of interest before the propagation}

\item{gf_prop}{sample's profile of interest after the first step of propagation with the gene-gene interaction network}

\item{ann_net_b}{matrix, for each row presents the gene identifier, the chromosome in which the gene is, the starting end ending position in the sequence.}

\item{frag_pattern}{string, initial character of the fragments name (e.g. "F")}

\item{ff_net}{edge list matrix of the gene fragment interaction network}

\item{ff_prop}{sample's profile of interest after the second step of propagation with the gene-fragment interaction network}
}
\value{
igraph object
}
\description{
Given the input and output of a two step network based propagation, it assembles an igraph object with all
the information included in order to be visualized or analysed
}
