% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_RWR.R
\name{diag_prof}
\alias{diag_prof}
\title{Convert matrix to diagonal matrix of row elements}
\usage{
diag_prof(input_m)
}
\arguments{
\item{input_m}{Matrix}
}
\value{
Return a diagonal matrix of the row elements
}
\description{
Convert matrix to diagonal matrix of row elements
}
