% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ML.R
\name{enhancer_classifier}
\alias{enhancer_classifier}
\title{Enhancer classifier based on centrality measures and propagation}
\usage{
enhancer_classifier(
  info,
  proj_name = "enh_pred",
  annot = 2,
  n_cores = 2,
  perf_measure = "classif.acc",
  tune_perf = 0.7,
  n_folds = 5
)
}
\arguments{
\item{info}{Info dataframe containing centrality measures and propagation values about the fragments and their corresponding nodes}

\item{proj_name}{name fo the project as character string}

\item{annot}{Define the minimum number of annotations to define an enhancer (e.g. 5)}

\item{n_cores}{Define number of cores to use in order to parallelize the tune, train and test tasks}

\item{perf_measure}{Set performace measure (e.g. classif.acc)}

\item{tune_perf}{Performance to reach in the tuning phase (e.g 0.7 of accuracy)}

\item{n_folds}{Number of folds for the cross validation (e.g. 5)}
}
\value{
List of 7 elements: proj_name, perfs_df, data_l, task, instance, lrn_tuned, lrn_tested, best_lrn
proj_name is the name of the project and run
perfs_df is the dataframe containing the classification performances
data_l is the list containing the original data processed by the classifier
task is the mlr3 object containing the task of classification
instance is the mlr3 object containing the tune, train and test parameters
lrn_tuned is the mlr3 object containing the classifier after the tuning
lrn_tested is the mlr3 object containing the classifier after the testing
best_lrn is the mlr3 object containing the classifier which best classifiers enhancers
}
\description{
This function creates a classifier able to predict if a node is an enhancer or not based
on centrality measures and propagation values obtained with the two step propagation.
First, it creates the task, tunes a ranger random forest, trains the model and predict
a randomly selected testing set in cross validation.
}
