% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ML.R
\name{explain_classifier}
\alias{explain_classifier}
\title{Explain the Enhancer classifier}
\usage{
explain_classifier(res_ml, n_cores = 2)
}
\arguments{
\item{res_ml}{Output list produced and obtained by running the function: enhancer_classifier}

\item{n_cores}{Define number of cores to use in order to parallelize the tune, train and test tasks}
}
\value{
List of 8 elements: task_dalex, lrn_dalex, ranger_exp, fi_ranger, fi_ranger_df, bd_ranger_enh, bd_ranger_no, best_lrn, pr_ranger
task_dalex is the dalex object containing the task of classification
lrn_dalex is the dalex object containing the learner
fi_ranger is the dalex object containing the trained and tested classifier
fi_ranger_df is the dataframe indicating how much each node's feature contributed in the classification
bd_ranger_enh is the dalex object containing the classification of an enhancer
bd_ranger_no is the dalex object containing the classification of a non enhancer
}
\description{
This function offers an explanation about how the classifier predicts the class of a
node/fragment and shows how each node's feature is contribuing in the classification.
}
