% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ML.R
\name{firstup}
\alias{firstup}
\title{Make the first letter a capital letter}
\usage{
firstup(x)
}
\arguments{
\item{x}{character string}
}
\value{
Make the first letter a capital letter
}
\description{
This function makes the first letter a capital letter while all the other letters low
}
