% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_GUI.R
\name{fun2}
\alias{fun2}
\title{Assign colours to node types in the igraph object}
\usage{
fun2(subnet)
}
\arguments{
\item{subnet}{igraph object}
}
\description{
Support function for the network visualization
}
