% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_RWR.R
\name{get_contrib_propXdest}
\alias{get_contrib_propXdest}
\title{Given the target nodes of the propagation, it provides the seed nodes of each target}
\usage{
get_contrib_propXdest(ff_prop, no_cores = 1)
}
\arguments{
\item{ff_prop}{numeric matrix of propagated profiles: seed nodes at the columns and target nodes at the rows}

\item{no_cores}{Default 1, An integer value greater than 0 indicating the number of cores to use for computing in parallel this task}
}
\value{
A list with contr_prop_l: For each target node as element of the list, the vector of the seed nodes which gave information to it
Plus, also the original propagation profiles of the target nodes.
}
\description{
It takes the result of a run of network-based propagation and determines which and how much information
the seed nodes gave to a specific target node
}
