% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_RWR.R
\name{get_contrib_propXorig}
\alias{get_contrib_propXorig}
\title{Given the seed nodes of the propagation, it provides the target nodes of each seed}
\usage{
get_contrib_propXorig(ff_prop, no_cores = 1)
}
\arguments{
\item{ff_prop}{Matrix of propagated profiles: seed nodes at the columns and target nodes at the rows}

\item{no_cores}{Default 1, An integer value greater than 0 indicating the number of cores to use for parallel computing}
}
\value{
A list with contr_prop_l: For each seed node as element of the list, the vector of the targets nodes which receive information from it
Plus, also the original propagation profiles of the seed nodes.
}
\description{
It takes the result of a run of network-based propagation and determines which and how much information
the target nodes received from a specific seed node
}
