% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_RWR.R
\name{get_dist_m}
\alias{get_dist_m}
\title{Determines the shortest distances from all the genes to the non genic fragments}
\usage{
get_dist_m(gf_net, ff_net, frag_pattern = "Frag")
}
\arguments{
\item{gf_net}{Edge list representing a network such that first column are genes and second column are "FragX" fragments}

\item{ff_net}{Edge list representing a network such that first and second column are "FragX" fragments}

\item{frag_pattern}{String pattern to identify a fragment by its nomenclature in the edge list (e.g. "Frag")}
}
\value{
Distance matrix of genes and non-genic fragments
}
\description{
Given the gene-genic fragment network and the fragment-fragment network, it computes the distances from the genes to the non genic fragments
}
