% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_RWR.R
\name{get_single_prop_data}
\alias{get_single_prop_data}
\title{Supportive function to infer the networks and the profiles needed in the single gene propagation}
\usage{
get_single_prop_data(
  gf_net,
  ff_net,
  input_m,
  dist_m,
  gene_in,
  frag_pattern = "Frag",
  degree = 2
)
}
\arguments{
\item{gf_net}{Edge list representing a network such that first column are genes and second column are "FragX" fragments}

\item{ff_net}{Edge list representing a network such that first and second column are "FragX" fragments}

\item{input_m}{Numeric matrix of cell profiles}

\item{dist_m}{Distance matrix computed between the genes and the non-genic fragments in the network}

\item{gene_in}{Character vector of the genes of interest}

\item{frag_pattern}{Character string to identify the fragments in the edge list (e.g. "FragX")}

\item{degree}{Default 2, the degree of distance which is used to select the most important fragments connected to the genes of interest}
}
\value{
A list with elements describing the input data which are by the method
}
\description{
Supportive function to infer the networks and the profiles needed in the single gene propagation
}
