% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ML.R
\name{merge_prop_info}
\alias{merge_prop_info}
\title{Merge propagation results with meta information}
\usage{
merge_prop_info(ff_prop, info)
}
\arguments{
\item{ff_prop}{Propagation result obtained with the second step of the multi-gene two-step propagation}

\item{info}{dataframe, first column is character vector of nodes, second colum is numeric vector of number of enhancer annotations associated to a node,
third column is the binarized version of the second column (1 the node has enhancer annotations, 0 otherwise), the
left columns are the centrality measures}
}
\value{
New info dataframe containing the input information combined to the propagation values
}
\description{
This function merges propagation results with meta information providing the number of enhancer
annotions associated to the fragments in the network
}
