% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ML.R
\name{range01}
\alias{range01}
\title{Zero - one standardization}
\usage{
range01(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
0-1 standardized vector
}
\description{
This function performs 0-1 standardization on a number vector
}
