% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ML.R
\name{rank_scale}
\alias{rank_scale}
\title{Maximum-based ranking normalization}
\usage{
rank_scale(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
Normalized vector based on max rank
}
\description{
This function performs Maximum-based ranking normalization on a number vector
}
