% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_RWR.R
\name{rwr_OVprop}
\alias{rwr_OVprop}
\title{Network based propagation based on Random walk with restart}
\usage{
rwr_OVprop(
  g,
  input_m,
  norm = "column",
  no_cores = 2,
  r = 0.8,
  stop_step = 200,
  stop_delta = 1e-06,
  keep_zero = F
)
}
\arguments{
\item{g}{Feature interaction network as adjacency matrix}

\item{input_m}{Numeric matrix of sample profiles}

\item{norm}{Default column, options: row or column or laplacian, it indicates the graph normalization performed on the network}

\item{no_cores}{Default 2, An integer value greater than 0 indicating the number of cores to use for parallel computing}

\item{r}{Default 0.8, A double value lower than 1 indicating the percentage of information that a gene keeps (0.8 is 80 percentage)}

\item{stop_step}{Default 200, An integer value greater than 0 indicating the number of iterations of the propagation}

\item{stop_delta}{Default 1e-06, A double value lower than 0 indicating the threshold under which all imputed propagation values are set 0}

\item{keep_zero}{Default FALSE, Boolean: TRUE if to keep profiles without information}
}
\value{
Propaged sample's profiles
}
\description{
This function performs propagation of the sample's profiles passed as columns of a matrix.
The sample's profile is characterized by the expression or aboundance of gene or fragments also called generically features
of the sample into account.
}
