% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_RWR.R
\name{rwr_SGprop}
\alias{rwr_SGprop}
\title{Single gene network based propagation based on random walk with restart}
\usage{
rwr_SGprop(
  gf_net,
  ff_net,
  gene_in,
  input_m,
  frag_pattern = "frag",
  out_rda = "sg_prop.rda",
  degree = 4,
  r1 = 0.1,
  r2 = 0.8,
  no_cores = 2
)
}
\arguments{
\item{gf_net}{Edge list of the chromatin interaction network such that first column are genes and second column are "FragX" fragments}

\item{ff_net}{Edge list of the chromatin interaction network such that first and second column are "FragX" fragments}

\item{gene_in}{Character vector of the genes of interest}

\item{input_m}{numeric matrix of cell profile, one column of a cell, rows are genes or fragments, values are expression or IAS}

\item{frag_pattern}{Character string to identify the fragments in the edge list (e.g. "FragX")}

\item{out_rda}{Default: sg_prop.rda, Character string to define the output rda file which will contain all the results of the analysis and intermediate products}

\item{degree}{Default 4, the degree of distance which is used to select the most important fragments connected to the genes of interest (e.g. degree 2 means G --> F1 --> F2)}

\item{r1}{Default 0.1, A double value lower than 1 indicating the percentage of information that a gene keeps (0.1 is 10 percentage)}

\item{r2}{Default 0.8, A double value lower than 1 indicating the percentage of information that a fragment keeps (0.8 is 80 percentage)}

\item{no_cores}{Default 2, An integer value greater than 0 indicating the number of cores to use for parallel computing}
}
\value{
A list single_gene_prop composed of two sublists
contr_lxDest: For each target node as element of the list, the vector of the seed nodes which gave information to it
contr_lxOrig: For each seed node as element of the list, the vector of the targets nodes which receive information from it
}
\description{
Performs the single gene propagation based on Random walk with restart.
This operation is suggested to apply only after having performed the standard propagation based on all the genes with the two steps.
First, overall propagation to diffuse expression of all genes included in a cell profile to genic fragments.
Second, overall propagation to diffuse the propagated expression of all genes from genic fragments to intergenic.
}
\details{
This function performs the propagation of individual genes of interest belonging to a cell's expression profile.
It does not perform the standard propagation using all the genes of a cell profile.
It then returns how much each gene of interest contributed to give information to the fragments.
It then returns how much each fragment received information from the genes of interest.
This function helps to understand the contribution of the individual genes in the two-step standard propagation.
}
