% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_GUI.R
\name{start_GUI}
\alias{start_GUI}
\title{Starts a shiny gui for the visualization of propagation results}
\usage{
start_GUI(net1, ann_net_b)
}
\arguments{
\item{net1}{igraph object to visualize}

\item{ann_net_b}{dataframe, specifies the genes positions (chromosome, start and end position)}
}
\description{
Given the graph obtained with create_net2plot function, the annotation of the genes included in the graph with
download_genes_annotation and the length of the chromosomes, it starts a graphical user interface to visualize
the propagation values on the network including genes and fragments.
This is useful to visualize the results of the two step propagation or the single gene propagation.
}
