% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_GUI.R
\name{toolkitf}
\alias{toolkitf}
\title{Connects network nodes to metainformation}
\usage{
toolkitf(net1, ann_net_b, frag_pattern = "F", ff_net = NULL)
}
\arguments{
\item{net1}{igraph object}

\item{ann_net_b}{annotation dataset of the nodes included in the igraph object}

\item{frag_pattern}{character string to identify the fragment nodes by their name in the igraph object}

\item{ff_net}{edge list of the gene fragment network}
}
\description{
Support function for the network visualization
}
