% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ML.R
\name{tuning_prop_vars}
\alias{tuning_prop_vars}
\title{Get the optimal isolation parameters for the multi-gene two-step propagation}
\usage{
tuning_prop_vars(gf_net, ff_net, input_m, info = NULL, n_cores = 2)
}
\arguments{
\item{gf_net}{Edge list of the chromatin interaction network such that first column are genes and second column are "FragX" fragments}

\item{ff_net}{Edge list of the chromatin interaction network such that first and second column are "FragX" fragments}

\item{input_m}{numeric matrix of cell profile, one column of a cell, rows are genes or fragments, values are expression or IAS}

\item{info}{2-column dataframe, first column is character vector of node names, second colum is numeric vector of number of enhancer annotations associated to a node}

\item{n_cores}{number of cores to run the function with parallel computing}
}
\value{
List of two elements: combinations and best_comb
combinations contains all the pair of values tested
best_comb contains only the optimal two values of the isolation paramter for the first and second overall propagation
}
\description{
This function determines the optimal values for the isolation parameter in the two
step propagation, one value for the propagation with the gene-fragment network and one
with the fragment-fragment network
}
