% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ML.R
\name{get_centr_info}
\alias{get_centr_info}
\title{Get centrality measures about nodes}
\usage{
get_centr_info(gf_net, ff_net, info)
}
\arguments{
\item{gf_net}{Edge list of the chromatin interaction network such that first column are genes and second column are "FragX" fragments}

\item{ff_net}{Edge list of the chromatin interaction network such that first and second column are "FragX" fragments}

\item{info}{2-column dataframe, first column is character vector of node names (genes and fragments), second colum is numeric vector of number of enhancer annotations associated to a node}
}
\value{
The info matrix with centrality measures associated to the nodes as new meta information
}
\description{
This function determines the centrality measures of the genes and fragments included in the
CIN and that have a prior information about being or not being enhancer
}
