% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_RWR.R
\name{rwr_OVprop}
\alias{rwr_OVprop}
\title{Network based propagation with random walk for the multi-gene two-step propagation}
\usage{
rwr_OVprop(
  g,
  input_m,
  norm = "column",
  no_cores = 2,
  r = 0.8,
  stop_step = 200,
  stop_delta = 1e-06,
  keep_zero = F
)
}
\arguments{
\item{g}{igraph, edge list or adjacency matrix of the chromatin interaction network (CIN)}

\item{input_m}{numeric matrix of cell profiles, columns are cells, rows are genes or fragments, value is expression of IAS}

\item{norm}{Default column, options: row or column or laplacian, it sets the graph normalization performed on the network}

\item{no_cores}{Default 2, An integer value greater than 0 indicating the number of cores to use for computing in parallel the task}

\item{r}{Default 0.8, A double value lower than 1 indicating the percentage of information that a gene keeps (0.8 is 80 percentage)}

\item{stop_step}{Default 200, An integer value greater than 0 indicating the number of iterations of the propagation}

\item{stop_delta}{Default 1e-06, A double value lower than 0 indicating the threshold under which all imputed propagation values are set 0}

\item{keep_zero}{Default FALSE, Boolean: TRUE if to keep profiles without information}
}
\value{
matrix of propagated profiles with imputed activity scores for the genes and fragments in the network and that replace the original expression profiles
}
\description{
This function performs propagation of each individual cell expression profile associated to a column of an expression matrix.
The cell expression profile is characterized by the expression of genes at the rows.
The values of the genes indicate their expression in a specific column and cell profile
The propagation considers each cell profile at a time. It maps the expression values to the corresponding genes represented as nodes in the network.
It diffuses the gene expression values. Each node in the CIN network composed by genes, genic fragments and intergenic fragments gets a propagation score which
measures its activity. The imputed activity score (IAS) obtained by genes and fragments replaces their original value in the cell expression profile.
The result is a new matrix of cell profiles. Each cell profile is now described with genes and also fragments who own a IAS.
}
