% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_RWR.R
\name{rwr_SGprop}
\alias{rwr_SGprop}
\title{Performs the single gene network based propagation}
\usage{
rwr_SGprop(
  gf_net,
  ff_net,
  gene_in,
  frag_pattern = "frag",
  out_rda = "sg_prop.rda",
  degree = 4,
  r1 = 0.1,
  r2 = 0.8,
  no_cores = 2
)
}
\arguments{
\item{gf_net}{Edge list representing a network such that first column are genes and second column are "FragX" fragments}

\item{ff_net}{Edge list representing a network such that first and second column are "FragX" fragments}

\item{gene_in}{Character vector of the genes of interest}

\item{frag_pattern}{Character string to identify the fragments in the edge list (e.g. "FragX")}

\item{out_rda}{Default: sg_prop.rda, Character string to define the output rda file which will contain all the results of the analysis and intermediate products}

\item{degree}{Default 4, the degree of distance which is used to select the most important fragments connected to the genes of interest (e.g. degree 2 means G --> F1 --> F2)}

\item{r1}{Default 0.1, A double value lower than 1 indicating the percentage of information that a gene keeps (0.1 is 10 percentage)}

\item{r2}{Default 0.8, A double value lower than 1 indicating the percentage of information that a fragment keeps (0.8 is 80 percentage)}

\item{no_cores}{Default 2, An integer value greater than 0 indicating the number of cores to use for parallel computing}
}
\value{
A list single_gene_prop composed of two sublists
contr_lxDest: For each target node as element of the list, the vector of the seed nodes which gave information to it
contr_lxOrig: For each seed node as element of the list, the vector of the targets nodes which receive information from it
}
\description{
Performs the single gene network based propagation based on Random walk with restart and
determines their contribution in the standard overall propagation
This function performs the propagation of individual genes of the starting sample's profile.
It does not perform the standard propagation using all the sample's genes.
The result is a propagated profile associate to each desired gene which was present in the initial sample's profile
and at the same time present in the input network.
Then it determines how much a gene contributed to give information to the fragments.
Then it determines how much a fragment received information from the genes
}
