% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_ML.R
\name{tuning_prop_vars}
\alias{tuning_prop_vars}
\title{Get the optimal isolation parameters for the two step propagation}
\usage{
tuning_prop_vars(gf_net, ff_net, input_m, info = NULL, n_cores = 2)
}
\arguments{
\item{gf_net}{Edge list of a network composed by genes and genetic fragments}

\item{ff_net}{Edge list of a network composed only by fragment interactions}

\item{input_m}{Numeric matrix of sample profiles}

\item{info}{2-column dataframe, first column is character vector of nodes, second colum is numeric vector of number of enhancer annotations associated to a node}

\item{n_cores}{number of cores to run the function with parallel computing}
}
\value{
List of two elements: combinations and best_comb
combinations contains all the pair of values tested
best_comb contains only the optimal two values of the isolation paramter for the first and second overall propagation
}
\description{
This function determines the optimal values for the isolation parameter in the two
step propagation, one value for the propagation with the gene-fragment network and one
with the fragment-fragment network
}
